:art: Fixing Field Types

+++ database/migrations/2020_07_09_204932_create_carts_table.php
@@ -15,7 +15,7 @@ class CreateCartsTable extends Migration
     {
         Schema::create('carts', function (Blueprint $table) {
             $table->id();
-            $table->double('total_price')->default(0.0);
+            $table->decimal('total_price', 8, 2)->default(0.0);
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->timestamps();
         });
